+++ src/main/java/ru/ryauzov/testtask/services/ClientService.java
@@ -9,5 +9,5 @@ public interface ClientService {
     void create(Client client);
     void update(Client client);
     void delete(Client client);
-    Client getById(int id);
+    Client getById(long id);
 }
